Keep <br>s - breaks existing localStorage TODO text
<br>s are now no longer replaced with newlines, so we don't have to use
whitesoace: pre and long lines will be allowed to wrap.

This commit introduces some weirdness - when leading and trailing linebreaks are
present and are trimmed off before TODO text state is updated, the linebreaks
aren't removed from display until the TODO is moved into another section or the
page is refreshed.